\name{computeMSE}
\alias{computeMSE}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Predictive accuracy estimates (MSE) across trees for linear or poisson regression model.
}
\description{
Predictive accuracy estimates (MSE) across trees for linear or poisson regression model. 
}
\usage{
computeMSE(response, predictions)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{response}{
a vector of actual response of outcome variable.
}
  \item{predictions}{
a vector of predicted response for the same outcome variable.
}
}
\value{
MSE estimates
}
