% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mob_fit_checksplit}
\alias{mob_fit_checksplit}
\title{Utility Function. Taken from party package to remove ":::" warning}
\usage{
mob_fit_checksplit(split, weights, minsplit)
}
\arguments{
\item{split}{see party package}

\item{weights}{see party package}

\item{minsplit}{see party package}
}
\description{
Utility Function. Taken from party package to remove ":::" warning
}
