% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mob_fit_childweights}
\alias{mob_fit_childweights}
\title{Utility Function. Taken from party package to remove ":::" warning}
\usage{
mob_fit_childweights(node, mf, weights)
}
\arguments{
\item{node}{see party package}

\item{mf}{see party package}

\item{weights}{see party package}
}
\description{
Utility Function. Taken from party package to remove ":::" warning
}
