% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mob_fit_fluctests}
\alias{mob_fit_fluctests}
\title{Utility Function. Taken from party package to remove ":::" warning}
\usage{
mob_fit_fluctests(obj, mf, minsplit, trim, breakties, parm)
}
\arguments{
\item{obj}{see party package}

\item{mf}{see party package}

\item{minsplit}{see party package}

\item{trim}{see party package}

\item{breakties}{see party package}

\item{parm}{cheese?}
}
\description{
Utility Function. Taken from party package to remove ":::" warning
}
