% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mob_fit_getlevels}
\alias{mob_fit_getlevels}
\title{Utility Function. Taken from party package to remove ":::" warning}
\usage{
mob_fit_getlevels(x)
}
\arguments{
\item{x}{see party package}
}
\description{
Utility Function. Taken from party package to remove ":::" warning
}
