% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mob_fit_getobjfun}
\alias{mob_fit_getobjfun}
\title{Utility Function. Taken from party package to remove ":::" warning}
\usage{
mob_fit_getobjfun(obj, mf, weights, left, objfun = deviance)
}
\arguments{
\item{obj}{see party package}

\item{mf}{see party package}

\item{weights}{see party package}

\item{left}{see party package}

\item{objfun}{see party package}
}
\description{
Utility Function. Taken from party package to remove ":::" warning
}
