% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mob_fit_setupnode}
\alias{mob_fit_setupnode}
\title{Utility Function. Taken from party package to remove ":::" warning}
\usage{
mob_fit_setupnode(obj, mf, weights, control)
}
\arguments{
\item{obj}{see party package}

\item{mf}{see party package}

\item{weights}{see party package}

\item{control}{see party package}
}
\description{
Utility Function. Taken from party package to remove ":::" warning
}
