% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mob_fit_splitnode}
\alias{mob_fit_splitnode}
\title{Utility Function. Taken from party package to remove ":::" warning}
\usage{
mob_fit_splitnode(x, obj, mf, weights, minsplit, objfun, verbose = TRUE)
}
\arguments{
\item{x}{see party package}

\item{obj}{see party package}

\item{mf}{see party package}

\item{weights}{see party package}

\item{minsplit}{see party package}

\item{objfun}{see party package}

\item{verbose}{To print or not to print}
}
\description{
Utility Function. Taken from party package to remove ":::" warning
}
