% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclass.R
\docType{class}
\name{mobforest.control-class}
\alias{mobforest.control-class}
\title{Class \code{"mobforest.control"} of mobForest model}
\description{
Control parameters for random forest
}
\section{Objects from the Class}{
 Objects can be created by
\code{\link{mobforest.control}}.
}

\examples{

# showClass("mobforest.control") The following code creates following forest
# controls: 400 trees to be constructed, sampling with replacement, a node
# contains at least 200 observations
mobforest_controls = mobforest.control(ntree = 400, mtry = 4,
    replace = TRUE, minsplit = 200)

}
\keyword{classes}
