% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobforest.output.R
\name{mobforest.output}
\alias{mobforest.output}
\title{Model-based random forest object}
\usage{
mobforest.output(oob_predictions, general_predictions,
  new_data_predictions, varimp_object, model_used, family, train_response,
  new_response = data.frame(matrix(0, 0, 0)))
}
\arguments{
\item{oob_predictions}{Predictions on out-of-bag data.}

\item{general_predictions}{Predictions on learning data.}

\item{new_data_predictions}{Predictions on new test data.}

\item{varimp_object}{The variable importance object.}

\item{model_used}{The model used.}

\item{family}{A description of the error distribution and link function to be
used in the model.}

\item{train_response}{Response outcome of training data.}

\item{new_response}{Response outcome of test data.}
}
\description{
Random Forest Output object that stores all the results including
predictions, variable importance matrix, model, family of error
distributions, and observed responses.
}
\seealso{
\code{\linkS4class{prediction.output}},
\code{\linkS4class{varimp.output}}
}
