% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclass.R
\docType{class}
\name{prediction.output-class}
\alias{prediction.output-class}
\title{Class \code{"prediction.output"} of mobForest model}
\description{
The object of this class stores predictions and predictive accuracy
estimates.
}
\section{Objects from the Class}{
 Objects can be created by calls of the
form \code{\link{prediction.output}}.
}

\seealso{
\code{\link{prediction.output}}, \code{\link{predictive.acc}}
}
\keyword{classes}
