% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.output.R
\name{prediction.output}
\alias{prediction.output}
\title{Predictions and predictive accuracy estimates}
\usage{
prediction.output(pred_mean = numeric(), pred_sd = numeric(),
  residual = numeric(), R2_or_acc = numeric(), mse = numeric(),
  overall_r2_or_acc = numeric(), pred_type = character())
}
\arguments{
\item{pred_mean}{Mean predictions across trees.}

\item{pred_sd}{Standard deviation predictions across trees.}

\item{residual}{Residuals (predicted outcome - observed outcome).}

\item{R2_or_acc}{R2 or accuracy (for binomial) across trees}

\item{mse}{MSE across trees}

\item{overall_r2_or_acc}{Overall R2 or accuracy (for binomial)}

\item{pred_type}{Out-of-bag data or test data or learning data.}
}
\value{
An object of class \code{"\link[=prediction.output]{prediction.output()}"}.
}
\description{
This function takes predictions and predictive accuracy estimates as input
arguments and creates objects of class
\code{\linkS4class{prediction.output}}.
}
\seealso{
\code{\linkS4class{prediction.output}},
\code{\link{mobforest.analysis}}
}
