\name{stringFormula}
\alias{stringFormula}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
formula converted to character format
}
\description{
Model in the formula object converted to a character character
}
\usage{
stringFormula(formula)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{formula}{
formula object
}
}

\value{
character. model
}
