% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclass.R
\docType{class}
\name{varimp.output-class}
\alias{varimp.output-class}
\title{Class \code{"varimp.output"} of mobforest model}
\description{
Variable importance
}
\section{Objects from the Class}{
 Objects can be created by calls of the
form \code{\link{varimp.output}}. %% ~~ describe objects here ~~
}

\references{
Strobl, C., Malley, J. and Tutz, G. (2009) An introduction to
recursive partitioning: rationale, application, and characteristics of
classification and regression trees, bagging, and random forests,
\emph{Psychol Methods}, 14, 323-348.\cr
}
\seealso{
\code{\linkS4class{varimp.output}}
}
