% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimp.output.R
\name{varimp.output}
\alias{varimp.output}
\title{Variable importance matrix containing the decrease in predictive accuracy
after permuting the variables across all trees}
\usage{
varimp.output(varimp_matrix)
}
\arguments{
\item{varimp_matrix}{a matrix containing decrease in predictive accuracy for
all variables for each tree}
}
\value{
An object of class \code{\linkS4class{varimp.output}}.
}
\description{
Values of variable 'm' in the oob cases are randomly permuted and R2
obtained through variable-m-permuted oob data is subtracted from R2
obtained on untouched oob data. The average of this number over all the
trees in the forest is the raw importance score for variable m.
}
\references{
Strobl, C., Malley, J. and Tutz, G. (2009) An introduction to
recursive partitioning: rationale, application, and characteristics of
classification and regression trees, bagging, and random forests,
\emph{Psychol Methods}, 14, 323-348.\cr
}
