\name{varimpOutput-class}
\Rdversion{1.1}
\docType{class}
\alias{varimpOutput-class}

\title{Class \code{"varimpOutput"}}
\description{
 Variable importance
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{\link{varimp_output}}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{varimpMatrix}:}{a matrix containing raw importance scores for all the variables per tree }
  }
}
\section{Methods}{
No methods defined with class "varimpOutput" in the signature.
}
\references{
Strobl, C., Malley, J. and Tutz, G. (2009) An introduction to recursive partitioning: rationale, application, and characteristics of classification and regression trees, bagging, and random forests, \emph{Psychol Methods}, 14, 323-348.\cr
}

\seealso{
\code{\linkS4class{varimpOutput}}
}

